refactor SOLR_select into a new file
cache/retry wrapper for json.loads(urllib2.urlopen) will go here
